fix: treat input spent as non-canonical
This commit change the behavior when input is flagged as spent. Previously, it would be omitted while now it
would be consider non-canonical instead. Omitting would lead to edge case that some input would be locked to
the system while it should be able to exit.

Also, this makes changes to how we flag inputs/outputs:
1) flag only when finalizing the exit
2) for canonical mark all inputs spend and exited outputs spend
3) for non-canonical mark only exiting inputs

issue: https://github.com/omisego/plasma-contracts/issues/470